refactor:CardUtils로 유틸성 메소드 추출
